Move common functions into shared DocumentUtils class
